# Optimize submodules fetching