@kreisys 2021-11-04
... siphoning submodules stakeholders to https://github.com/NixOS/nix/pull/5497