@Kha 2023-03-18
Closing in favor of #6530, see also https://github.com/NixOS/nix/pull/7862#issuecomment-1474858948